Finalisation components && réorganisation fonctions profil
