+++ app/components/faq/faqQuestion.tsx
@@ -1,0 +1,9 @@
+export default function Question ({question, answer}:{question:any, answer:any}) {
+
+    return (
+        <div className="faq-question">
+          <h2>{question}</h2>
+          <p>{answer}</p>
+        </div>
+    )
+}
