+++ app/components/markdown/markdown.tsx
@@ -11,7 +11,7 @@ type MarkdownRendererProps = {
 };
 
 const MarkdownRenderer: FunctionComponent<MarkdownRendererProps> = ({ source }) => {
-  // console.log('source == ', source);
+
   return (
     <>
       <ReactMarkdown
