+++ app/components/usersAdminWidget/userCard.tsx
@@ -1,5 +1,4 @@
 import React from 'react'
-import axios from 'axios'
 import { useState } from 'react';
 import BanModal from './banConfirmationModal';
 
@@ -13,8 +12,8 @@ export default function userCard({person}:any) {
   return (
     <div>
       <div className='user-card-div'>
-          <h1>{person.uuid}</h1>
-          <button onClick={handleClick} className='ban-button'>Ban User</button>
+        <h1>{person.uuid}</h1>
+        <button onClick={handleClick} className='ban-button'>Ban User</button>
       </div>
       {openModal && <BanModal person={person} closeModal={setOpenModal}/>}
     </div>
